+++ src/pint/models/solar_system_shapiro.py
@@ -1,9 +1,9 @@
 """Solar system Shapiro delay."""
-import logging
 
 import astropy.constants as const
 import astropy.units as u
 import numpy
+from loguru import logger as log
 
 from pint import (
     Tearth,
@@ -18,8 +18,6 @@ from pint import (
 )
 from pint.models.parameter import boolParameter
 from pint.models.timing_model import DelayComponent
-
-log = logging.getLogger(__name__)
 
 
 class SolarSystemShapiro(DelayComponent):